+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/EICAS/ECAM/A320_Neo_LowerECAM_CRZ.js
@@ -1,3 +1,21 @@
+/*
+ * A32NX
+ * Copyright (C) 2020-2021 FlyByWire Simulations and its contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 /** @type A320_Neo_LowerECAM_CRZ */
 var A320_Neo_LowerECAM_CRZ;
 (function (A320_Neo_LowerECAM_CRZ) {
@@ -22,16 +40,10 @@ var A320_Neo_LowerECAM_CRZ;
             TemplateElement.call(this, this.init.bind(this));
         }
         init() {
-            this.isInMetric = BaseAirliners.unitIsMetric(Aircraft.A320_NEO);
+            this.unitConversion = parseFloat(NXDataStore.get("CONFIG_USING_METRIC_UNIT", "1"));
 
             // Set units
-            if (this.isInMetric) {
-                this.querySelector("#FuelUsedUnit").textContent = "KG";
-                this.unitFactor = SimVar.GetSimVarValue("FUEL WEIGHT PER GALLON", "kilogram");
-            } else {
-                this.querySelector("#FuelUsedUnit").textContent = "LBS";
-                this.unitFactor = SimVar.GetSimVarValue("FUEL WEIGHT PER GALLON", "lbs");
-            }
+            this.querySelector("#FuelUsedUnit").textContent = this.unitConversion === 1 ? "KG" : "LBS";
 
             // Get value fields
             this.FuelUsedTotal = this.querySelector("#FuelUsedTotal");
@@ -119,17 +131,18 @@ var A320_Neo_LowerECAM_CRZ;
             }
 
             // Fuel
-            const leftConsumption = this.isInMetric ?
-                SimVar.GetSimVarValue("GENERAL ENG FUEL USED SINCE START:1", "KG") :
-                SimVar.GetSimVarValue("GENERAL ENG FUEL USED SINCE START:1", "gallon") * this.unitFactor * 0.001;
-            const rightConsumption = this.isInMetric ?
-                SimVar.GetSimVarValue("GENERAL ENG FUEL USED SINCE START:2", "KG") :
-                SimVar.GetSimVarValue("GENERAL ENG FUEL USED SINCE START:2", "gallon") * this.unitFactor * 0.001;
+            const leftConsumption = SimVar.GetSimVarValue("GENERAL ENG FUEL USED SINCE START:1", "KG") * this.unitConversion;
+            const rightConsumption = SimVar.GetSimVarValue("GENERAL ENG FUEL USED SINCE START:2", "KG") * this.unitConversion;
 
             const leftConsumptionShown = fastToFixed(leftConsumption - (leftConsumption % 10), 0);
             const rightConsumptionShown = fastToFixed(rightConsumption - (rightConsumption % 10), 0);
 
-            if ((leftConsumptionShown != this.LeftConsumptionDisplayed) || (rightConsumptionShown != this.RightConsumptionDisplayed)) {
+            if ((leftConsumptionShown !== this.LeftConsumptionDisplayed) || (rightConsumptionShown !== this.RightConsumptionDisplayed)) {
+                const unitConversion = parseFloat(NXDataStore.get("CONFIG_USING_METRIC_UNIT", "1"));
+                if (unitConversion !== this.unitConversion) {
+                    this.unitConversion = unitConversion;
+                    this.querySelector("#FuelUsedUnit").textContent = this.unitConversion === 1 ? "KG" : "LBS";
+                }
                 const totalConsumptionShown = parseInt(leftConsumptionShown) + parseInt(rightConsumptionShown);
                 this.FuelUsedTotal.textContent = totalConsumptionShown;
                 this.FuelUsedLeft.textContent = leftConsumptionShown;