+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/EICAS/ECAM/A320_Neo_LowerECAM_Engine.js
@@ -1,3 +1,21 @@
+/*
+ * A32NX
+ * Copyright (C) 2020-2021 FlyByWire Simulations and its contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 var A320_Neo_LowerECAM_Engine;
 (function (A320_Neo_LowerECAM_Engine) {
     class Definitions {
@@ -30,8 +48,10 @@ var A320_Neo_LowerECAM_Engine;
             TemplateElement.call(this, this.init.bind(this));
         }
         init() {
-            this.engineLeft = new EngineInfo(1, this.querySelector("#LeftGauges"), this.querySelector("#FuelUsedValueLeft"), this.querySelector("#OilTemperatureValueLeft"), this.querySelector("#EngineBleedPressureValueLeft"), this.querySelector("#StartValveLeft_OPEN"), this.querySelector("#StartValveLeft_CLOSED"), this.querySelector("#N1VibrationLevelValueLeft"), this.querySelector("#N2VibrationLevelValueLeft"));
-            this.engineRight = new EngineInfo(2, this.querySelector("#RightGauges"), this.querySelector("#FuelUsedValueRight"), this.querySelector("#OilTemperatureValueRight"), this.querySelector("#EngineBleedPressureValueRight"), this.querySelector("#StartValveRight_OPEN"), this.querySelector("#StartValveRight_CLOSED"), this.querySelector("#N1VibrationLevelValueRight"), this.querySelector("#N2VibrationLevelValueRight"));
+            this.unitConversion = parseFloat(NXDataStore.get("CONFIG_USING_METRIC_UNIT", "1"));
+            this.querySelector("#FuelUsedUnit").textContent = this.unitConversion === 1 ? "KG" : "LBS";
+            this.engineLeft = new EngineInfo(1, this.querySelector("#LeftGauges"), this.querySelector("#FuelUsedValueLeft"), this.unitConversion, this.querySelector("#OilTemperatureValueLeft"), this.querySelector("#EngineBleedPressureValueLeft"), this.querySelector("#StartValveLeft_OPEN"), this.querySelector("#StartValveLeft_CLOSED"), this.querySelector("#N1VibrationLevelValueLeft"), this.querySelector("#N2VibrationLevelValueLeft"));
+            this.engineRight = new EngineInfo(2, this.querySelector("#RightGauges"), this.querySelector("#FuelUsedValueRight"), this.unitConversion, this.querySelector("#OilTemperatureValueRight"), this.querySelector("#EngineBleedPressureValueRight"), this.querySelector("#StartValveRight_OPEN"), this.querySelector("#StartValveRight_CLOSED"), this.querySelector("#N1VibrationLevelValueRight"), this.querySelector("#N2VibrationLevelValueRight"));
             this.ignTitleText = this.querySelector("#IGNTitle");
             this.ignLeftValueText = this.querySelector("#IGNValueLeft");
             this.ignRightValueText = this.querySelector("#IGNValueRight");
@@ -43,11 +63,16 @@ var A320_Neo_LowerECAM_Engine;
             if (!this.isInitialised || !A320_Neo_EICAS.isOnBottomScreen()) {
                 return;
             }
+            const unitConversion = parseFloat(NXDataStore.get("CONFIG_USING_METRIC_UNIT", "1"));
+            if (this.unitConversion !== unitConversion) {
+                this.unitConversion = unitConversion;
+                this.querySelector("#FuelUsedUnit").textContent = this.unitConversion === 1 ? "KG" : "LBS";
+            }
             if (this.engineLeft != null) {
-                this.engineLeft.update(_deltaTime);
+                this.engineLeft.update(_deltaTime, this.unitConversion);
             }
             if (this.engineRight != null) {
-                this.engineRight.update(_deltaTime);
+                this.engineRight.update(_deltaTime, this.unitConversion);
             }
             this.updateIGN();
         }
@@ -135,9 +160,10 @@ var A320_Neo_LowerECAM_Engine;
     }
     A320_Neo_LowerECAM_Engine.Page = Page;
     class EngineInfo {
-        constructor(_engineIndex, _gaugeDiv, _fuelUsedValueText, _oilTemperatureValueText, _engineBleedPressureValueText, _startValveOpenLine, _startValveClosedLine, _N1VibrationValueText, _N2VibrationValueText) {
+        constructor(_engineIndex, _gaugeDiv, _fuelUsedValueText, _unitConversion, _oilTemperatureValueText, _engineBleedPressureValueText, _startValveOpenLine, _startValveClosedLine, _N1VibrationValueText, _N2VibrationValueText) {
             this.engineIndex = _engineIndex;
             this.fuelUsedValueText = _fuelUsedValueText;
+            this.unitConversion = _unitConversion;
             this.oilTemperatureValueText = _oilTemperatureValueText;
             this.engineBleedPressureValueText = _engineBleedPressureValueText;
             this.engineBleedValveOpenLine = _startValveOpenLine;
@@ -182,8 +208,9 @@ var A320_Neo_LowerECAM_Engine;
             this.setN1VibrationValue(0, true);
             this.setN2VibrationValue(0, true);
         }
-        update(_deltaTime) {
-            this.setFuelUsedValue(SimVar.GetSimVarValue("GENERAL ENG FUEL USED SINCE START:" + this.engineIndex, "kg"));
+        update(_deltaTime, _unitConversion) {
+            this.unitConversion = _unitConversion;
+            this.setFuelUsedValue(SimVar.GetSimVarValue("GENERAL ENG FUEL USED SINCE START:" + this.engineIndex, "kg") * this.unitConversion);
             this.setOilTemperatureValue(SimVar.GetSimVarValue("GENERAL ENG OIL TEMPERATURE:" + this.engineIndex, "celsius"));
 
             let bleedPressure = 0;
@@ -213,7 +240,7 @@ var A320_Neo_LowerECAM_Engine;
             return value;
         }
         setFuelUsedValue(_value, _force = false) {
-            if ((_value != this.fuelUsedValue) || _force) {
+            if ((_value !== this.fuelUsedValue) || _force) {
                 this.fuelUsedValue = _value;
                 if (this.fuelUsedValueText != null) {
                     this.fuelUsedValueText.textContent = fastToFixed(this.fuelUsedValue, 0);