+++ .github/CHANGELOG.md
@@ -20,6 +20,7 @@
 1. [FLIGHTMODEL] Improved fix for flaps 1 unrecoverable dives - @donstim (donbikes#4084)
 1. [ECAM/CORE] Improved flap auto-retraction detection - @MisterChocker (Leon)
 1. [MCDU] Enabled buttons on the F/O MCDU - @tyler58546 (tyler58546)
+1. [ECAM] Added missing unit conversion to cruise and engine page - @MisterChocker (Leon)
 
 ## 0.5.1
 1. [CDU] Allow SimBrief user IDs as well as usernames - @pareil6 (pareil6)